Move commands from test runner into a separate folder
